requirements.txt
asgiref==3.8.1
Django==4.2.4
django-cors-headers==4.6.0
django-safedelete==1.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Markdown==3.4.4
mypy==1.6.1
mypy-extensions==1.0.0
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2024.2
sqlparse==0.5.3
typing_extensions==4.12.2
